+++ Views/Shared/_Layout.cshtml
@@ -17,7 +17,7 @@
             <div id="logo">
                 <h1><span class="icon icon-cog"></span><a href="#">ReptiHurt</a></h1>
                 <span>Jesteś zalogowany: @Html.Raw(HttpContextAccessor.HttpContext.Request.Cookies["userLogged"])</span>
-                <button type="button" class="myButton">@Html.ActionLink("Wyloguj", "LogOut", "Home")</button>
+                <button type="button" class="myButton">@Html.ActionLink("Wyloguj", "LogOut", "Account")</button>
                 <button class="myButton">@Html.ActionLink("Koszyk", "Cart", "Cart", new { OrderId = ViewBag.OrderId })</button>
                 <nav>
                     <ul>
