+++ Views/Shared/_LayoutAdmin.cshtml
@@ -17,7 +17,7 @@
             <div id="logo">
                 <h1><span class="icon icon-cog"></span><a href="#">ReptiHurt</a></h1>
                 <span>Jesteś zalogowany: @Html.Raw(HttpContextAccessor.HttpContext.Request.Cookies["userLogged"])</span>
-                <button class="myButton">@Html.ActionLink("Wyloguj", "LogOut", "Home")</button>
+                <button class="myButton">@Html.ActionLink("Wyloguj", "LogOut", "Account")</button>
                 <nav>
                     <ul>
                         <li>
