+++ Views/Shared/_LayoutTerms.cshtml
@@ -18,7 +18,7 @@
             <div id="logo">
                 <h1><span class="icon icon-cog"></span><a href="#">ReptiHurt</a></h1>
                 <span>Jesteś zalogowany: @Html.Raw(HttpContextAccessor.HttpContext.Request.Cookies["userLogged"])</span>
-                <button class="myButton">@Html.ActionLink("Wyloguj", "LogOut", "Home")</button>
+                <button class="myButton">@Html.ActionLink("Wyloguj", "LogOut", "Account")</button>
                 <button class="myButton">
                     @Html.ActionLink("Koszyk", "Cart", "Cart", new { OrderId = ViewBag.OrderId })
                 </button>
